rgw: bucket link: base "bucket move" (tenant id only)
This is the base or primitive "bucket move" function.  It handles rewriting
the endpoint and info rados objects for the rgw bucket, plus deleting
the original rados objects that are no longer correct.  It doesn't
handle changing the bucket name; that's in a future commit.  Some of
the changes here will get overwritten by that commit.

Fixes: http://tracker.ceph.com/issues/35885

Signed-off-by: Marcus Watts <mwatts@redhat.com>