mgr/prometheus:add disk metadata to occupation metric
This patch exposes the devices and device_ids meta
data from the osd to the ceph_disk_occupation
metric. This potentially allows osd alerts to merge
with the occupation data to present a disk failure
message that points to a specific device/disk serial.

AWS, vmware and baremetal all populate both device_ids
and devices metadata - but in the event either of these
is empty we default to 'N/A' so at least the label is
consistent.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>