mon: add an "osd force_healthy_stretch_mode" command
The requirements to automatically exit recovery stretch mode are stringent:
no degraded, inactive, or unknown PGs.
If you know that you can peer immediately and don't want to wait for that
recovery process to finish, or for some reason recovery can't complete
but you are okay with the availability risk, you can issue this command
to shortcut it.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>