mon: osd: switch crush_bucket_mandatory_member to be CRUSH_ITEM_NONE-based
Previously we compared it to zero, but we could technically want to require
osd.0 as a member, maybe? In any case we have a "DNE" indicator
in CRUSH_ITEM_NONE, so use it.

Also, for osd_types.h, declare a pg_pool_t::pg_CRUSH_ITEM_NONE to use,
since apparently we can't import crush.h there and hard-coding it is bad.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>