client: clean up the unnecessary client_lock for _conf->client_trace
There is no need to make the "cct->_conf->client_trace" under the
client_lock, it is for the "ceph-syn" and it almost won't change,
even it will but it still won't cause any race.

For the "inode->ino" and "inode->snapid", since they won't change
once initialized when creating the inode, and the inode here will
be protected by the reference count, so no need the client_lock too.

Fixes: https://tracker.ceph.com/issues/46768
Signed-off-by: Xiubo Li <xiubli@redhat.com>