os/bluestore: eradicate bufferlist rebuilds around FileWriter
This commits reworks the way how `FileWriter::flush_buffer()`
complements the buffer with zeros when the requested flush
length is not a multiply of block size.

Originally, it was zeroing the `buffer` directly -- without
an assistance from `page_aligned_appender` though which all
regular appends flow. This mismatch was fragmenting `buffer`
and leading to rebuilds in e.g. `KernelDevice::aio_write`.

In the patch, `page_aligned_appender` gets `append_zero()`
and `FileWriter` directs the filling-with-zeros through
the appender, and thus allowing it to put the content
contiguously in memory -- without creating a new `ptr_node`.

Credits for detecting the problem belong to Jianpeng Ma
<jianpeng.ma@intel.com>. Thanks!

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>