doc/cephadm: managing ceph.conf
This commit repairs PR#36796, which was about
updating the configuration settings in ceph.conf
without touching the file directly.

That commit wasn't building because of a couple of
back quotes that were touching the word that followed
them. So much like the princess from The Princess and
The Pea is Sphinx. So careful must we always be.

Fixes: https://tracker.ceph.com/issues/47265
Signed-off-by: Zac Dover <zac.dover@gmail.com>