librbd: don't attempt to read from object snapshots known to not exist
If there is an issue with the list-snaps command that results in a forced
read of the full object, don't attempt to read from any snapshots that
are known to not exist. This can occur if cache-tiering does not have the
full snapshot context or if there is a bug in the list-snaps op
implementation.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>