SubmittingPatches-backports.rst: clarify where to explain
The prevailing consensus is that rationales of direct fixes need
not be included in the commit messages. The matter must be
made clear to the reviewers, but it's not important how this is
accomplished.

(Perusers of the git history interested in finding out this information
are free to search for it by linking the commit in question to the PR it
originated from, and possibly from there to the related tracker issues.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>