qa/suite/rados: Introduce "scheduler" directory within rados perf suite.
Introduce a "scheduler" directory under the rados:perf tree to allow perf
suite to specify tests with the default scheduler(WPQ) and also with
the dmClock scheduler. One specification also overrides the number of
shards(1) and the number of threads per shard(16) to test with apart from
the default settings. This allows testing and performance benchmarking
with the new proposal to use one shard and multiple threads per shard with
the dmClock scheduler.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>