msg: use timeout when initiating connection
There's a timer that enforces a connection timeout but it isn't scoped
properly. It's started only after the ``connect`` request completes.
Because of that, connections can hang in STATE_CONNECTING_RE state
indefinitely if the "connect" operation doesn't complete.

This change will start the timer when entering the STATE_CONNECTING
state.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>