rgw: address crash and race in RGWIndexCompletionManager
An atomic int was used in a modulo operator to distribute contention
among a set of locks and to track completions. Because it was an int,
enough increments would cause it to go negative (due to
twos-complement encoding and overflow) thereby causing a
crash. Additionally, even though it was atomic, the read and increment
were separate operations, leading to a race.

This commit addresses both of these issues.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#2080433
(cherry picked from commit 9f20d0292c5f7b341bc2bc580fb119416323fdc1)
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>