mgr/dashboard: configure rbd mirroring
One-click button in the case of an orch cluster for configuring the
rbd-mirroring when its not properly setup. This button will create an
rbd-mirror service and also an rbd labelled pool(replicated: size-3) (if they are not
existing)

Resolves: rhbz#2053470
Fixes: https://tracker.ceph.com/issues/55646
Signed-off-by: Nizamudeen A <nia@redhat.com>

 Conflicts:
	src/pybind/mgr/dashboard/frontend/src/app/core/error/error.component.html
	src/pybind/mgr/dashboard/frontend/src/app/shared/services/module-status-guard.service.ts
	src/pybind/mgr/dashboard/tests/test_rbd_mirroring.py

error component and module-status-guard.service.ts had one minor
conflict which was resolved by getting incoming changes (add missing
logic).
test_rbd_mirroring had an import conflict which was resolved by
accepting incoming changes which had the same imports with new ones.