mgr/dashboard: Consider null values as zero in grafana panels
After upgrading from RHCS4 to RHCS5..some of the grafana charts broke.
This is because in RHCS5 we do not generate the metrics if its value is
zero as a result the null value from that metric breaks the grafana
charts or graphs. This PR is to fix the above mentioned issue.

Fixes: https://tracker.ceph.com/issues/63088
Resolves: rhbz#2250826

Signed-off-by: Aashish Sharma <aasharma@redhat.com>
(cherry picked from commit 6f3f58cb8e4ce100cc7186858465b4d11d5c2c49)